requirements.txt
flask
gunicorn
psycopg[binary]
python-dotenv
flask-limiter
email-validator
